+++ HISTORY.rst
@@ -2,6 +2,12 @@
 History
 =======
 
+2.1.4 (2017-04-06)
+------------------
+    * Card #215: new script to export all transition history of Orders to a tsv file (rudaporto).
+    * Card #218: default value for empty submission_path in the Assignment must be None (rudaporto).
+
+
 2.1.3 (2017-04-05)
 ------------------
     * Card #184: new script to add missing transitions to Order and Assignments using Ophelie's data set (rudaporto).